Add files via upload
Lately I have been using the Time System Library which comes with ADRIFT 5. Besides %Minute% and %Hour% it also has a variable called %Day% which I think is intended to keep track of which day it is. You could use this to print the date, e.g. if we are somewhere between day 4 and day 30 in the game, you could output "Date: %Day%th of July".

Unfortunately, it turns out, that the variable %Day% is not correctly updated when time passes midnight. For instance, it may take on negative values. And it is not updated when you sleep.

I have now modified the time system so that the variable %Day% is now correctly updated, both at midnight and when you sleep. Thus you can now use the Improved Time System Library to keep track of the date in the game.