Version 3 is supplied both as a taf-file and as an amf-file.
Version 3 released. The zip-file includes both a taf-file and an amf-file. Unless you are very confident in switching between libraries, it is recommended to use the taf-file whenever you create a new game and to have all libraries deselected at all times. Otherwise, there is a risk, that you overwrite the library in an existing taf-file by accident.